[main@4b0de08] Update AL-Go System Files from microsoft/AL-Go-PTE@preview - e747532 / Related to AB#539394 (#318)
## preview

Note that when using the preview version of AL-Go for GitHub, we
recommend you Update your AL-Go system files, as soon as possible when
informed that an update is available.

### AL-Go Telemetry

AL-Go now offers a dataexplorer dashboard to get started with AL-Go
telemetry. Additionally, we've updated the documentation to include a
couple of kusto queries if you would rather build your own reports.

### Issues

- Issue 1770 Wrong type of _projects_ setting in settings schema
- Issue 1787: Publish to Environment from PR fails in private repos

### Add custom jobs to AL-Go workflows

It is now possible to add custom jobs to AL-Go workflows. The Custom Job
needs to be named `CustomJob<something>` and should be placed after all
other jobs in the .yaml file. The order of which jobs are executed is
determined by the Needs statements. Your custom job will be executed
after all jobs specified in the Needs clause in your job and if you need
the job to be executed before other jobs, you should add the job name in
the Needs clause of that job. See
[https://aka.ms/algosettings#customjobs](https://aka.ms/algosettings#customjobs)
for details.

Note that custom jobs might break by future changes to AL-Go for GitHub
workflows. If you have customizations to AL-Go for GitHub workflows, you
should always doublecheck the pull request generated by Update AL-Go
System Files.

### Support for Custom AL-Go template repositories

Create an AL-Go for GitHub repository based on
[https://aka.ms/algopte](https://aka.ms/algopte) or
[https://aka.ms/algoappsource](https://aka.ms/algoappsource), add custom
workflows, custom jobs and/or settings to this repository and then use
that repository as the template repository for other repositories. Using
custom template repositories allows you to create and use highly
customized template repositories and control the uptake of this in all
repositories. See
[https://aka.ms/algosettings#customtemplate](https://aka.ms/algosettings#customtemplate)
for details.

> [!NOTE]
> Customized repositories might break by future changes to AL-Go for
GitHub. If you are customizing AL-Go for GitHub, you should always
double-check the pull request when updating AL-Go system files in your
custom template repositories.

Related to AB#539394

Co-authored-by: mazhelez <mazhelez@users.noreply.github.com>