[main@d3fe45e] Update AL-Go System Files from microsoft/AL-Go-PTE@preview - cb1099b / Related to AB#539394 (#339)
## preview

Note that when using the preview version of AL-Go for GitHub, we
recommend you Update your AL-Go system files, as soon as possible when
informed that an update is available.

### Mechanism to overwrite complex settings type

By default, AL-Go merges settings from various places (see [settings
levels](https://aka.ms/algosettings#where-are-the-settings-located)).
Basic setting types such as `string` and `integer` are overwritten, but
settings with complex types such as `array` and `object` are merged.

However, sometimes it is useful to avoid merging complex types. This can
be achieved by specifying `overwriteSettings` property on a settings
object. The purpose of the property is to list settings, for which the
value will be overwritten, instead of merged. Read more at
[overwriteSettings
property](https://aka.ms/algosettings#overwriteSettings)

### AL Code Analysis tracked in GitHub

AL-Go already supports AL code analysis, but up until now this was not
tracked in GitHub. It is now possible to track code analysis issues
automatically in the GitHub security tab, as well as having any new
issues posted as a comment in Pull Requests.

Enable this feature by using the new setting
[trackALAlertsInGithub](https://aka.ms/algosettings#trackALAlertsInGithub).
This setting must be enabled at the repo level, but can optionally be
disabled per project.

Please note that some automated features are premium and require the use
of [GitHub Code
Security](https://docs.github.com/en/get-started/learning-about-github/about-github-advanced-security)

### Issues

- Discussion 1885 Conditional settings for CI/CD are not applied
- Discussion 1899 Remove optional properties from "required" list in
settings.schema.json

Related to AB#539394

Co-authored-by: mazhelez <mazhelez@users.noreply.github.com>